Initial ebuild for osl
OSL is highly dependent upon the correct versions of boost and llvm

OSL 1.5.12 failed to build. It is reported to work with LLVM 3.4 and Boost 1.55.0 under Ubuntu
but I was not able to compile LLVM-3.4 on my system (install fails when it attepts to copy
a file onto itself). It does not work under LLVM 3.5

OSL 1.6.9+ gave an undefined reference error compiling against Boost-1.61.0,
which was resolved by using Boost 1.55,0 1.56,0 and 1.60.0, although
a missing symbol for ObjectBuffer in LLVMMCJIT prevented successful linkage.

By backporting a patch from the master branch to RB-1.7 and Release it is possible
to remove the dependency on ObjectBuffer and compile osl-1.7.2 and 1.7.3 under LLVM-3.5
Apparently LLVM-3.4 might also work but this is ~amd64 and fails to compile on my system
Original patch at https://github.com/imageworks/OpenShadingLanguage/commit/77edb88eba19e0899da40e2d91ee04aaa3f0cb42

I have tested with Blender 2.77a which finds and uses 1.7.2. 1.7.3 and the master branch successfully.

TODO:
The doc use flag is ignored - documentation is installed regardless
The documentation is installed in /usr/doc rather than /usr/share/doc/OSL
The test use flag is restricted as the tests fail with this patch (see original patch comment)