osl: Release 1.7.2 with patchset
1.7.3 was a live ebuild which can never be make stable. We can get
around that by having the official release 1.7.2 patched with select
patches from RB-1.7 branch. Made a patchset in the OpenShadingLanguage
repon on my github under the branch of RB-1.7-Gentoo. That can be
rebased to upstream and new patches against Release-1.7.2 tag. I
merged in the two patches for fix docs and remove mcjit to make sure
the patch would cleanly apply with no offsets.

Added in ~ARCH keywords.