blender: Add warning about removing cuda kernel and change to dodoc
Since the CUDA kernel is compiled outside of portage, it's not tracked
so add this warning so the user may be interested in removing it.

Also, if using CUDA, OpenCL gets in the way, so set a required use to
not use OpenCL when CUDA is enabled. OpenCL is slower than CUDA and
no point in using it ever even have it installed.

Apparently, dohtml has been depreciated in EAPI 6, so switch to
dodoc and add in the html directory to the dointo var.