Add support for OSL with modern LLVM
OSL upstream contains a pull request #693 from marsupial which is
included in osl-build-with-modern-llvm.patch. This allows OSL to be
built with LLVM 3.9.

I resurrected one of the old OSL pre-ebuilds and updated it to use
this patch. I also needed to add -DLLVM_STATIC=ON to force it to link
with the LLVM shared libraries that gentoo provides.

Finally I have added the osl USE flag back to blender.

With this setup, I am now able to render using OSL shaders.